add TestAlgorithmicStepMethods with some tests & implementation of some algorithmic step methods in Thrust
add pairs to backend parametrisation
fixed test_floor_out_of_place (dtype of output)
add generate_is_first_in_pair

+++ PySDM_tests/unit_tests/backends/test_maths_methods.py
@@ -46,7 +46,7 @@ class TestMathsMethods:
     @staticmethod
     def test_floor_out_of_place(sut, shape_2d):
         params = [{'name': "output",
-                   'details': {'shape': shape_2d, 'dtype': float}},
+                   'details': {'shape': shape_2d, 'dtype': int}},
                   {'name': "input_data",
                    'details': {'shape': shape_2d, 'dtype': float, 'negative': True}},
                   ]
